lms-knowledge-graph/requirements.txt
neo4j==6.1.0
pytz==2026.1.post1
python-dotenv==1.0.1
